feat: add middleware handler wrappers
